Refactor names
Refactored names that were causing issues
