+++ backend/src/main/java/me/magicbudget/controller/UserController.java
@@ -1,9 +1,9 @@
 package me.magicbudget.controller;
 
 
-import me.magicbudget.dto.incoming_request.LoginUserRequest;
-import me.magicbudget.dto.incoming_request.RegistrationAndAuthRequest;
-import me.magicbudget.dto.outgoing_response.LoginUserResponse;
+import me.magicbudget.dto.incomingrequest.LoginUserRequest;
+import me.magicbudget.dto.incomingrequest.RegistrationAndAuthRequest;
+import me.magicbudget.dto.outgoingresponse.LoginUserResponse;
 import me.magicbudget.model.UserInformation;
 import me.magicbudget.security.service.RegistrationAndAuthService;
 import me.magicbudget.service.UserService;
