+++ backend/src/test/java/me/magicbudget/controller/IncomeControllerTest.java
@@ -2,9 +2,9 @@ package me.magicbudget.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.transaction.Transactional;
-import me.magicbudget.dto.incoming_request.IncomeRequest;
-import me.magicbudget.dto.incoming_request.LoginUserRequest;
-import me.magicbudget.dto.incoming_request.RegistrationAndAuthRequest;
+import me.magicbudget.dto.incomingrequest.IncomeRequest;
+import me.magicbudget.dto.incomingrequest.LoginUserRequest;
+import me.magicbudget.dto.incomingrequest.RegistrationAndAuthRequest;
 import me.magicbudget.model.Income;
 import me.magicbudget.model.IncomeType;
 import me.magicbudget.model.User;
