+++ backend/src/test/java/me/magicbudget/controller/SavingGoalControllerTest.java
@@ -9,8 +9,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import org.springframework.http.MediaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.transaction.Transactional;
-import me.magicbudget.dto.incoming_request.LoginUserRequest;
-import me.magicbudget.dto.incoming_request.RegistrationAndAuthRequest;
+import me.magicbudget.dto.incomingrequest.LoginUserRequest;
+import me.magicbudget.dto.incomingrequest.RegistrationAndAuthRequest;
 import me.magicbudget.model.SavingGoal;
 import me.magicbudget.model.UserInformation;
 import me.magicbudget.model.User;
